Move round files to the rounds folder
